* all commit messages:
  small comment update in bgjobs
  settings log update
  replaced tools.sendmail with django email sending
  Set file-loglevel to info
  adminx index page 2 cols in grid
  Added enabled field to AdminValue model
  AdminValues doc update
  Readme update
  RebuildFlowFrontpage bgtask
  mainapp/bgjobs.py: Background Jobs with Threading
  Cache ajax page loads
  adminvalues: override settings with the admin interface
  Start AdminValues: helper to override settings/defaults with AdminValues
  admin interface updates; bugfix in models.Photo.post_delete (check whether file exists before removing)
  Templates use minified js + css files
  build.py minify css + js
  nginx config tweak: Cache-Control public for images
  Minor fix: in photoflow set header-width margin -13 to compensate for flow right margin
  Bugfix: photoflow ajax load more with >1 block works now. Adjusted flow-initial to 3 blocks, perpage to 2 blocks to reduce loadings

+++ app/mainapp/models.py
@@ -54,6 +54,7 @@ class AdminValue(caching.base.CachingMixin, models.Model):
 
     key = models.CharField(max_length=50)
     val = models.TextField()
+    enabled = models.BooleanField(default=True)
 
     def __unicode__(self):
         return "<AdminValue(%s='%s')>" % (self.key, self.val)
@@ -217,11 +218,13 @@ class Photo(caching.base.CachingMixin, models.Model):
         photo = kwargs["instance"]
         log.info("Photo post delete handler: removing image files '%s'" % photo.local_filename)
         if photo.local_filename:
+            # Delete photos
             fn_upload = os.path.join(settings.MEDIA_ROOT, settings.MEDIA_DIR_UPLOAD, photo.local_filename)
             fn_photo = os.path.join(settings.MEDIA_ROOT, settings.MEDIA_DIR_PHOTOS, photo.local_filename)
-            os.remove(fn_upload)
-            os.remove(fn_photo)
-            log.info("- all removed")
+            if os.path.isfile(fn_upload):
+                os.remove(fn_upload)
+            if os.path.isfile(fn_photo):
+                os.remove(fn_photo)
         else:
             log.info("- no local filename")
